tests/wpt/tests/html/canvas/offscreen/drawing-rectangles-to-the-canvas/2d.strokeRect.basic.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.strokeRect.basic</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.strokeRect.basic</h1>
<p class="desc">strokeRect works</p>


<script>
var t = async_test("strokeRect works");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.strokeStyle = '#0f0';
  ctx.lineWidth = 50;
  ctx.strokeRect(25, 24, 50, 2);
  _assertPixel(canvas, 50,25, 0,255,0,255);
  t.done();

});
</script>
